Refactor App Layout into separate components
